requirements.txt
grpcio==1.66.2
grpcio-tools==1.66.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1